+++ daemon/src/devices/banglejsdevice.h
@@ -58,7 +58,7 @@ private:
 
     int m_infoBatteryLevel = 0;
     int m_steps = 0;
-    int m_heartrate = 0;
+    double m_heartrate = 0;
     QString m_firmwareVersion;
     QString m_hardwareVersion;
